+++ src/main/java/mod/gottsch/forge/treasure2/core/inventory/TreasureContainers.java
@@ -24,6 +24,7 @@ import mod.gottsch.forge.treasure2.core.setup.Registration;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraftforge.common.extensions.IForgeMenuType;
+import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.items.ItemStackHandler;
 import net.minecraftforge.registries.RegistryObject;
 
@@ -66,7 +67,7 @@ public class TreasureContainers {
 
 	}
 	
-	public static void register() {
-		Registration.registerContainers();
+	public static void register(IEventBus bus) {
+		Registration.registerContainers(bus);
 	}
 }