+++ CHANGELOG.md
@@ -9,9 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.1.0] - 2023-06-25
 
+### Added
+
+- Added Crate and Moldy Crate mimics.
+
 ### Changed
 
-- updated the topaz key texture to be more like the onyx key
+- Made 1.18.2 -> 1.192. porting code changes (ex . Random -> RandomSource)
+- Updated the topaz key texture to be more like the onyx key
+- Key durability is no longer setup during item registration but in common setup using Reflection (need to update 1.18.2)
 
 
 ## [3.0.0] - 2023-06-24